replace implementation module with declarative macro
The current approach of having a free standing module that is
instantiated multiple times with the explicit `path` often causes
confusion when reading the code. So replace this with a very simple
declarative macro that hopefully is more explicit and simpler than the
current approach.
