--- src/implementation/mod.rs
@@ -1,7 +1,0 @@
-// surfman/surfman/src/implementation/mod.rs
-//
-//! This is an included private module that automatically produces the implementations of the
-//! various traits for a backend.
-
-mod connection;
-mod device;
