+++ src/platform/windows/wgl/mod.rs
@@ -7,8 +7,7 @@ pub mod context;
 pub mod device;
 pub mod surface;
 
-#[path = "../../../implementation/mod.rs"]
-mod implementation;
+crate::implement_interfaces!();
 
 #[cfg(test)]
 #[path = "../../../tests.rs"]
